Samples: Split battery implementation
Extracted battery level calculation out of battery.c and refactored
names.
